refactor(ts-typings): restructure definitions  to separate files with only proper API exported
