+++ src/ts-typings/jsx.d.ts
@@ -1,98 +1,12 @@
-import { VDOMElement } from './index'
+import { HTMLProps, Attributes, ClassAttributes } from './common';
+import { VDOMElement, Component } from './api';
 export { }
-
-export type Key = string | number;
-export type Ref<T> = string | ((instance: T) => any);
-
-export type HTMLProps<T extends Element> = Partial<T> & Attributes<T>;
-
-export type Attributes<T> = IncrementalDomHTMLAttributes<T> & HyperscriptHTMLAttributes & HyperscriptEventHandler<T>;
-interface IncrementalDomHTMLAttributes<T> {
-  key?: Key,
-  ref?: Ref<T>,
-  statics?: string[],
-  skip?: boolean,
-}
-interface HyperscriptHTMLAttributes {
-  class?: string,
-}
-
-interface HyperscriptEventHandler<T> {
-  onAbort?: typeof HTMLElement.prototype.onabort;
-  onActivate?: typeof HTMLElement.prototype.onactivate;
-  onBeforeactivate?: typeof HTMLElement.prototype.onbeforeactivate;
-  onBeforecopy?: typeof HTMLElement.prototype.onbeforecopy;
-  onBeforecut?: typeof HTMLElement.prototype.onbeforecut;
-  onBeforedeactivate?: typeof HTMLElement.prototype.onbeforedeactivate;
-  onBeforepaste?: typeof HTMLElement.prototype.onbeforepaste;
-  onBlur?: typeof HTMLElement.prototype.onblur;
-  onCanplay?: typeof HTMLElement.prototype.oncanplay;
-  onCanplaythrough?: typeof HTMLElement.prototype.oncanplaythrough;
-  onChange?: typeof HTMLElement.prototype.onchange;
-  onClick?: typeof HTMLElement.prototype.onclick;
-  onContextmenu?: typeof HTMLElement.prototype.oncontextmenu;
-  onCopy?: typeof HTMLElement.prototype.oncopy;
-  onCuechange?: typeof HTMLElement.prototype.oncuechange;
-  onCut?: typeof HTMLElement.prototype.oncut;
-  onDblclick?: typeof HTMLElement.prototype.ondblclick;
-  onDeactivate?: typeof HTMLElement.prototype.ondeactivate;
-  onDrag?: typeof HTMLElement.prototype.ondrag;
-  onDragend?: typeof HTMLElement.prototype.ondragend;
-  onDragenter?: typeof HTMLElement.prototype.ondragenter;
-  onDragleave?: typeof HTMLElement.prototype.ondragleave;
-  onDragover?: typeof HTMLElement.prototype.ondragover;
-  onDragstart?: typeof HTMLElement.prototype.ondragstart;
-  onDrop?: typeof HTMLElement.prototype.ondrop;
-  onDurationchange?: typeof HTMLElement.prototype.ondurationchange;
-  onEmptied?: typeof HTMLElement.prototype.onemptied;
-  onEnded?: typeof HTMLElement.prototype.onended;
-  onError?: typeof HTMLElement.prototype.onerror;
-  onFocus?: typeof HTMLElement.prototype.onfocus;
-  onInput?: typeof HTMLElement.prototype.oninput;
-  onInvalid?: typeof HTMLElement.prototype.oninvalid;
-  onKeydown?: typeof HTMLElement.prototype.onkeydown;
-  onKeypress?: typeof HTMLElement.prototype.onkeypress;
-  onKeyup?: typeof HTMLElement.prototype.onkeyup;
-  onLoad?: typeof HTMLElement.prototype.onload;
-  onLoadeddata?: typeof HTMLElement.prototype.onloadeddata;
-  onLoadedmetadata?: typeof HTMLElement.prototype.onloadedmetadata;
-  onLoadstart?: typeof HTMLElement.prototype.onloadstart;
-  onMousedown?: typeof HTMLElement.prototype.onmousedown;
-  onMouseenter?: typeof HTMLElement.prototype.onmouseenter;
-  onMouseleave?: typeof HTMLElement.prototype.onmouseleave;
-  onMousemove?: typeof HTMLElement.prototype.onmousemove;
-  onMouseout?: typeof HTMLElement.prototype.onmouseout;
-  onMouseover?: typeof HTMLElement.prototype.onmouseover;
-  onMouseup?: typeof HTMLElement.prototype.onmouseup;
-  onMousewheel?: typeof HTMLElement.prototype.onmousewheel;
-  onMscontentzoom?: typeof HTMLElement.prototype.onmscontentzoom;
-  onMsmanipulationstatechanged?: typeof HTMLElement.prototype.onmsmanipulationstatechanged;
-  onPaste?: typeof HTMLElement.prototype.onpaste;
-  onPause?: typeof HTMLElement.prototype.onpause;
-  onPlay?: typeof HTMLElement.prototype.onplay;
-  onPlaying?: typeof HTMLElement.prototype.onplaying;
-  onProgress?: typeof HTMLElement.prototype.onprogress;
-  onRatechange?: typeof HTMLElement.prototype.onratechange;
-  onReset?: typeof HTMLElement.prototype.onreset;
-  onScroll?: typeof HTMLElement.prototype.onscroll;
-  onSeeked?: typeof HTMLElement.prototype.onseeked;
-  onSeeking?: typeof HTMLElement.prototype.onseeking;
-  onSelect?: typeof HTMLElement.prototype.onselect;
-  onSelectstart?: typeof HTMLElement.prototype.onselectstart;
-  onStalled?: typeof HTMLElement.prototype.onstalled;
-  onSubmit?: typeof HTMLElement.prototype.onsubmit;
-  onSuspend?: typeof HTMLElement.prototype.onsuspend;
-  onTimeupdate?: typeof HTMLElement.prototype.ontimeupdate;
-  onVolumechange?: typeof HTMLElement.prototype.onvolumechange;
-  onWaiting?: typeof HTMLElement.prototype.onwaiting;
-}
 
 declare global {
   // https://www.typescriptlang.org/docs/handbook/jsx.html
   namespace JSX {
-    interface ElementClass {
-    }
-
+    interface Element extends VDOMElement<any> { }
+    interface ElementClass extends Component<any> {}
     interface ElementAttributesProperty<Props> {
       // Special hack for own components type checking.
       // more detail, see: https://www.typescriptlang.org/docs/handbook/jsx.html
@@ -100,8 +14,8 @@ declare global {
       readonly _props: Props,
     }
 
-    interface Element extends VDOMElement<any> {
-    }
+    interface IntrinsicAttributes extends Attributes { }
+    interface IntrinsicClassAttributes<T> extends ClassAttributes<T> { }
 
     interface IntrinsicElements {
       a: HTMLProps<HTMLAnchorElement>;
